Fix Ignorable error detection/test
